move tests to dynamic op
